[SPARK-42944][PYTHON][FOLLOW-UP] Rename tests from foreachBatch to foreach_batch
### What changes were proposed in this pull request?

This PR proposes to rename tests from foreachBatch to foreach_batch.

### Why are the changes needed?

Non-API should follow snake_naming rule per PEP 8.

### Does this PR introduce _any_ user-facing change?

No, dev-only.

### How was this patch tested?

CI in this PR should test it out.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #42675 from HyukjinKwon/pyspark-connect.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
